# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: api-kit
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/admin/permissions:
    get:
      summary: 'Display a listing of the permissions.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: est
                      guard_name: web
                      roles:
                        -
                          id: 1
                          name: culpa
                          guard_name: web
                    -
                      id: 2
                      name: vitae
                      guard_name: web
                      roles:
                        -
                          id: 1
                          name: culpa
                          guard_name: web
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: est
                        guard_name: web
                        roles:
                          -
                            id: 1
                            name: culpa
                            guard_name: web
                      -
                        id: 2
                        name: vitae
                        guard_name: web
                        roles:
                          -
                            id: 1
                            name: culpa
                            guard_name: web
                    items:
                      type: object
      tags:
        - Permissions
    post:
      summary: 'Store a newly created permission in storage.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    name: est
                    guard_name: web
                    roles:
                      -
                        id: 1
                        name: culpa
                        guard_name: web
                properties:
                  data:
                    type: object
                    example:
                      id: 1
                      name: est
                      guard_name: web
                      roles:
                        -
                          id: 1
                          name: culpa
                          guard_name: web
      tags:
        - Permissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: ''
                  example: fugiat
                guard_name:
                  type: string
                  description: ''
                  example: est
                roles:
                  type: array
                  description: ''
                  example:
                    - et
                  items:
                    type: string
              required:
                - name
...
  /api/admin/roles:
    get:
      summary: 'Display a listing of the roles.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: culpa
                      guard_name: web
                      permissions:
                        -
                          id: 1
                          name: est
                          guard_name: web
                        -
                          id: 2
                          name: vitae
                          guard_name: web
                    -
                      id: 2
                      name: sit
                      guard_name: web
                      permissions:
                        -
                          id: 6
                          name: ut
                          guard_name: web
                        -
                          id: 7
                          name: earum
                          guard_name: web
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: culpa
                        guard_name: web
                        permissions:
                          -
                            id: 1
                            name: est
                            guard_name: web
                          -
                            id: 2
                            name: vitae
                            guard_name: web
                      -
                        id: 2
                        name: sit
                        guard_name: web
                        permissions:
                          -
                            id: 6
                            name: ut
                            guard_name: web
                          -
                            id: 7
                            name: earum
                            guard_name: web
                    items:
                      type: object
      tags:
        - Roles
    post:
      summary: 'Store a newly created role in storage.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    name: culpa
                    guard_name: web
                    permissions:
                      -
                        id: 1
                        name: est
                        guard_name: web
                      -
                        id: 2
                        name: vitae
                        guard_name: web
                properties:
                  data:
                    type: object
                    example:
                      id: 1
                      name: culpa
                      guard_name: web
                      permissions:
                        -
                          id: 1
                          name: est
                          guard_name: web
                        -
                          id: 2
                          name: vitae
                          guard_name: web
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: ''
                  example: qui
                guard_name:
                  type: string
                  description: ''
                  example: nulla
                permissions:
                  type: array
                  description: ''
                  example:
                    - quia
                  items:
                    type: string
              required:
                - name
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []